+++ packages/convai-widget-core/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @elevenlabs/convai-widget-core
 
+## 0.16.3
+
+### Patch Changes
+
+- Updated dependencies [d73eca4]
+  - @elevenlabs/client@1.20.0
+
+## 0.16.2
+
+### Patch Changes
+
+- Updated dependencies [a82a0c5]
+- Updated dependencies [118fd04]
+- Updated dependencies [f4c5588]
+  - @elevenlabs/client@1.19.1
+
+## 0.16.1
+
+### Patch Changes
+
+- Updated dependencies [880edd2]
+- Updated dependencies [a068144]
+  - @elevenlabs/client@1.19.0
+
 ## 0.16.0
 
 ### Minor Changes