+++ packages/convai-widget-embed/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elevenlabs/convai-widget-embed
 
+## 0.16.3
+
+## 0.16.2
+
+## 0.16.1
+
 ## 0.16.0
 
 ### Minor Changes