+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @elevenlabs/react-native
 
+## 1.2.22
+
+### Patch Changes
+
+- Updated dependencies [d73eca4]
+  - @elevenlabs/client@1.20.0
+  - @elevenlabs/react@1.12.4
+
+## 1.2.21
+
+### Patch Changes
+
+- Updated dependencies [a82a0c5]
+- Updated dependencies [118fd04]
+- Updated dependencies [f4c5588]
+  - @elevenlabs/client@1.19.1
+  - @elevenlabs/react@1.12.3
+
+## 1.2.20
+
+### Patch Changes
+
+- Updated dependencies [880edd2]
+- Updated dependencies [a068144]
+  - @elevenlabs/client@1.19.0
+  - @elevenlabs/react@1.12.2
+
 ## 1.2.19
 
 ### Patch Changes